// gameStopper/src/enums/color.ts
export enum ColorE {
  MAIN_BGD = "#2A2A2A",
  LIST_BGD = "#191919",
  LIST_ITEM_BGD = "#2D2D2D",
  LIST_ITEM_HOVERED_BGD = "#600202",
  LIST_ITEM_ACTIVE_BGD = "#9E0101",
  LIST_ITEM_SEPARATOR_BGD = "#420808",
  LIST_BORDER_COLOR = "#7E7E7E",
  TEXT_COLOR = "#FFFFFF",
  LIST_ITEM_FOCUSED_BGD = "#c90003",
}